Make usb boot target more flexible
Updating the makefile to support a variable for the drive letter you
want to make bootable. This saves people from adding every removable
drive letter to the makefile.

Added a specific target for the c: drive since that shouldn't ever be
written.

Really the shortcut target to make a bootable disk is pretty scary and
could probably be a utility instead of a makefile target.