[Gfx] ITS ALIVE!!!!!!!!!
Finally got the card out of D6 State and can read/write core registers!

Changed force wake to use the MT versions of the registers.

The first thing I've done now it's alive, is to set up the memory swizzle registers (needs to be done before first DRAM access).  The setting assume 2 DIMMS of same size.

Without the force wake, the values of these registers don't change.  The Tile Ctrl register is undocumented and comes from Linux driver, but it's value remains unmodified (0)... Maybe this is a legacy register?