Fix to shaders to make them work with the Intel assembler from: git://anongit.freedesktop.org/git/xorg/app/intel-gpu-tools
The open-source assembler uses a slightly different notation from
the documentation and what I am used too.

I've haven't got the assembler to compile on cygwin yet.
I had to do major hacking of the config file, just to get the
assembler to compile without complaining about missing dependacies for the
other tools on Linux.

The make_shaders.sh script can be used to build the shaders.
Shaders are compiled for Gen7 (IVB) with all warning enabled.
The -b flag causes the assembler to output in C style for including in the project
I've called these output's .c - though proberbly want to #include to reference them.

A single pixel shader has multiple entry point depending on the dispatch mode.
This is handle by using a different label for each entry point in the shader.
The *.ps_entry file is a list of these labels and is used by the assembler to do
apropreate padding of the different shaders.

The -e flag, generates #defines for the addresses of each of these entry points.
This is in the .h file.

Changes to the shaders themselves at this point is fairly minimal.
- The assembler uses 'g' notation for GRF registers instead of 'r'.
- It doesn't use the ":" before the register type value.
- Each line is seperated with ";"
- The assembler doesn't want type information on the source register for send instructions

Vertex Shader Changes:
- Assembler didn't like the width on the first move -- changed to 1, but need to check.

Pixel Shader Changes:
- I had an off-by 1 bug in the localtion of the Render Cache DP Op code in the send instructions.

I dissasembled the compiled code with:
./intel-gen4disasm -g 7 -b <shader.c>

There's still some things I need to review from the output.
- I can't remember what the 1Q/1H flags mean.  Need to sanity check.
- The code to set the channel enable mask is incorrect in the disasembly.
- Need to sanity check the reported send flags and sizes in the disaembly output
- Need to sanity check the vector float output in the diasembly
- Pixel shader sub register's are incorrect in the disasembly.